crate,parser: move conversion test #25
